:recycle:  [BE]refactor: mypage에 전해주는 dto에 팔로우 수, 팔로잉 수 포함

+++ back/munggle/src/main/java/com/munggle/follow/retpository/FollowRepository.java
@@ -15,7 +15,7 @@ public interface FollowRepository extends JpaRepository<Follow, FollowId> {
 
     List<Follow> findByFollowFromIdAndIsFollowedTrue(Long id);
 
-    Long countByFollowToIdAndIsFollowedTrue(Long id);
+    Integer countByFollowToIdAndIsFollowedTrue(Long id);
 
-    Long countByFollowFromIdAndIsFollowedTrue(Long userId);
+    Integer countByFollowFromIdAndIsFollowedTrue(Long userId);
 }
